Rework array updating and repainting logic
Previously, this used a timer to make repeated calls to an event listener. A better, clearer and more logical implementation would have been to just have an updating method that returned true or false based on the success of the update, and to call repaint() based on the return value of this method. Rewrite logic to use this implementation. Also add docstrings for init and update methods.